fix: Handle simulator battery level
This commit fixes a bug where the battery level was displayed as -100% on the simulator.

The expo-battery library returns -1 for the battery level on the simulator, which was being incorrectly converted to -100%.

This commit updates the useBatteryHelperViewModel to handle this case by setting the battery level to null. The BatteryHelperScreen is also updated to display 'Not available on simulator' when the battery level is null.